resolve duplicate logic
